services/wikiService/requirements.txt
pyzmq==27.1.0
wikipedia==1.4.0
python-dotenv==1.2.2